[core][presenter][refactor] remove: unnecessary base presenters
