--- moviper/src/main/java/com/mateuszkoslacz/moviper/base/presenter/ViperActivityBasePresenter.java
@@ -1,96 +1,0 @@
-package com.mateuszkoslacz.moviper.base.presenter;
-
-import android.os.Bundle;
-import android.support.annotation.NonNull;
-
-import com.hannesdorfmann.mosby.mvp.MvpBasePresenter;
-import com.mateuszkoslacz.moviper.iface.interactor.MoviperInteractor;
-import com.mateuszkoslacz.moviper.iface.presenter.MoviperPresenter;
-import com.mateuszkoslacz.moviper.iface.presenter.interactor.MoviperPresenterForInteractor;
-import com.mateuszkoslacz.moviper.iface.presenter.routing.MoviperPresenterForRouting;
-import com.mateuszkoslacz.moviper.iface.routing.MoviperRouting;
-import com.mateuszkoslacz.moviper.iface.view.ViperView;
-
-/**
- * Created by mateuszkoslacz on 08.08.2016.
- * <p>
- * Viper - View, Interactor, Presenter, Entities, Routing
- * <p>
- * This is a Activity version of base presenter class for mentioned set of concepts.
- * (see {@link MvpBasePresenter})
- * <p>
- * You can use any Mosby Activity View with this class
- * ({@link com.hannesdorfmann.mosby.mvp.MvpActivity},
- * {@link com.hannesdorfmann.mosby.mvp.lce.MvpLceActivity},
- * {@link com.hannesdorfmann.mosby.mvp.viewstate.MvpViewStateActivity},
- * {@link com.hannesdorfmann.mosby.mvp.viewstate.lce.MvpLceViewStateActivity})
- */
-//TODO migrate to MvpNullObjectPresenter base class?
-public abstract class ViperActivityBasePresenter
-        <ViewType extends ViperView,  // I prefer readability rather than conventions
-                InteractorType extends MoviperInteractor,
-                RoutingType extends MoviperRouting>
-        extends MoviperBasePresenter<ViewType>
-        implements MoviperPresenter<ViewType>,
-        MoviperPresenterForInteractor<InteractorType>,
-        MoviperPresenterForRouting<RoutingType> {
-
-    @NonNull
-    private RoutingType routing;
-
-    @NonNull
-    private InteractorType interactor;
-
-    public ViperActivityBasePresenter() {
-        super();
-        this.routing = createRouting();
-        this.interactor = createInteractor();
-    }
-
-    public ViperActivityBasePresenter(Bundle args) {
-        super(args);
-        this.routing = createRouting();
-        this.interactor = createInteractor();
-    }
-
-    @Override
-    @Deprecated
-    public boolean isRoutingAttached() {
-        return routing != null;
-    }
-
-    @Override
-    @Deprecated
-    public boolean isInteractorAttached() {
-        return interactor != null;
-    }
-
-    @Override
-    public void attachView(ViewType view) {
-        super.attachView(view);
-        //noinspection unchecked
-        routing.attachPresenter(this);
-        routing.attachActivity(view);
-        interactor.attachPresenter(this);
-    }
-
-    @Override
-    public void detachView(boolean retainInstance) {
-        super.detachView(retainInstance);
-        routing.detachPresenter();
-        routing.detachActivity();
-        interactor.detachPresenter();
-    }
-
-    @NonNull
-    @Override
-    public RoutingType getRouting() {
-        return routing;
-    }
-
-    @NonNull
-    @Override
-    public InteractorType getInteractor() {
-        return interactor;
-    }
-}
